Support replacing the underlying FS in go1.16
